Save built manpage to repo
The manpage needs to be built and committed as part of the release
process.

This is bad practice, but I really do not want to add Pandoc as a build
dependency, particularly in Arch.  An alternative is to write the
manpage directly, but I prefer using Markdown so that it can be viewed
on GitHub.